Title: 세 수의 합과 평균 구하기, Time: 132ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 250205/세 수의 합과 평균 구하기/sum-and-mean-of-three-numbers.java
@@ -7,6 +7,6 @@ public class Main {
 
         int a = sc.nextInt(), b = sc.nextInt(), c = sc.nextInt();
 
-        System.out.printf("%d %d", a + b + c, (a + b + c) / 3);
+        System.out.printf("%d %f", a + b + c, (double)(a + b + c) / 3);
     }
 }
